requirements.txt
numpy
pandas
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
speedtest-cli==2.1.3
tzdata==2023.3
